requirements.txt
antlr4-python3-runtime==4.9.3
certifi==2024.12.14
charset-normalizer==3.4.0
cloudpickle==3.1.0
Cython==3.0.11
demucs==4.0.1
diffq==0.2.4
dora_search==0.1.12
einops==0.8.0
filelock==3.16.1
filesystempro==2.0.0.0
fsspec==2024.10.0
idna==3.10
Jinja2==3.1.4
julius==0.2.7
lameenc==1.7.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
omegaconf==2.3.0
openunmix==1.3.0
psutil==6.1.0
PyYAML==6.0.2
requests==2.32.3
retrying==1.3.4
six==1.17.0
submitit==1.5.2
sympy==1.13.3
torch==2.2.2
torchaudio==2.2.2
tqdm==4.67.1
treetable==0.2.5
typing_extensions==4.12.2
urllib3==2.2.3
